Fix #1044: handle TypevarsMissContext mode in wildApprox
When `wildApprox` encounters a PolyParam it approximates it by its
bounds in the current constraint set, but this is incorrect if
`TypevarsMissContext` is set, we might get the bounds of another use of
this `PolyType`. This is exactly what happened in i1044.scala where the
implicit view `refArrayOps` needs to be used twice with two different
type parameters.

The fix is to approximate a PolyParam by its original bounds in its
PolyType if `TypevarsMissContext` is set.

This fix was inspired by the approach of #1054.