Refine validity checking
When defining a class in the interpreter we had a case where
the class was accessed at phase 46 in the backend, yet the denotation
was the initial denotation in a previous run. In that case we
have to check again at the phase where the denotation is valid.
This was not done before, and hence the owner of the denbotation
did not contain the symbol because the backend phase is after flatten.