Handle classOf in terms of ClassTag
Advantage: classOf is not magic anymore. It's also more
flexible as it gives us an easy way to go from a ClassTag
to a class.

For the moment we leave the Classof phase in, in order to
do the right thing if somebody references Predef.classOf
explicitly (our improved classOf is in DottyPredef). Once
we merge DottyPredef and Predef we can get rid of the phase.