Changes to owners in Mixin and Constructors
Two changes:

1. Replace changeOwer with changeOwnerAfter for code
that moves into the $initial methods in Mixin. This is needed because
otherwise subsequent transforms gets confused wrt new vs old owners.
`i1131.scala` exhibits the problem.

2. Drop `transformSym` changed the owner of tenplate-local symbols
to be the primary constructor. But that is done anyway with a "changeOwnerAfter"
in `intoConstr`. So it is redundant and actually gets in the way with
a `changeOwnerAfter` in `Mixin`. The faulty scenario is this:

  1. The SymTransformer of Constructor is run on a constructor-local definition.
     The owner of that definition is set to <init> after phase Constructors.

  2. The body of the definition is transformed in Mixin. The owner is set
     to the initializer method, but only for the interval between Mixin
     and Constructors.

Changing to changeOwner in Mixin avoided that problem by duplicating the symbol but
it runs into other problems.

Fortunately, the solution is much simpler than the status quo: Two changeOwnerAfter calls
and no SymTransformer.