Have checkNotPrivate skip over private aliases.
CheckNotPrivate now avoids private aliases, so that types sat prepresentable in
transformation phases.

This does not solve the problem that private classes or abstract types might leak coming
from Scala 2, but there is not really a good cure for that. We can reject them outright or
allow them under language:Scala2 and hope for the best.