Inline ArrayRuntime methods in ScalaRunTime.array_clone().
Support for Array[T].clone() was introduced in
13b1fcf89cb684fc81b5adfb6df68d3799263dda. At the time, all
calls to array.clone() were redirected to
ScalaRunTime.array_clone(), which meant that array_clone()
itself could not be implemented in terms of `x.clone()`. A
Java binding was necessary.

Now, the rewriting to `array_clone()` is only done for
unbounded generic arrays. Since all the calls within
`array_clone()` are monomorphic arrays, the rewriting is not
applied, and the Java binding is unnecessary.