Initial implementation featuring two different highlighters
One was implemted by hand and the other by using dotty's parser. The one
built by hand is shorter, and behaves correctly.

The scanner one is unfortunately not ready for testing - there are too
many things that are workarounds for it to be a good solution as of now

The code added from Ammonite is licensed under MIT, not sure where to
put the license - but will add it once I know.