partest: Enable separate compilation and javac tests
partest can separately compile files based on their suffix (_1, _2,
...), it turns out that this feature was never enabled in the dotty
version of partest and no one noticed (it prints warnings in
./tests/partest-generated/gen.log which no one reads), tests with *.java
files should be compiled both with javac and dotty, but compiling with
javac was also disabled.

Enabling this revealed some latent bugs that will be fixed in the next
few commits.