Fix #1292: give position when failing to emit switch on annotated match
This fix gives the position and refactors the code that gave off
warnings, but it also begs the question - should we be able to handle
emitting a switch for the following case:

```scala
(x: @switch) match {
  case 'a' if somePredicate(x) => // ...
  case 'b' => // ...
}
```

Currently if there is a guard, the patternmatcher will fail to emit
a switch. Scalac can handle these cases currently.