Eta-expand unapplied types that have type parameters
We would like to change from a scheme where eta-expansion was
prototype driven to one where unapplied parameterized types are always
eta expanded. The reason is that we might miss some eta expansions due
to cyclic references.

    run/colltest4

is an exmaple. Here, we missed an eta expansion in the type of Iterator.
The class definition is:

    trait Iterable[+A] extends IterableOnce[A] with FromIterable[Iterable] {

We'd expect that the second parent would expand to

    FromIterable[[X0] -> Iterable[X0]]

But we miss the expansion because at the time we complete Iterable we have not
completed FromIterable yet. In fact this happens in both the old and the new hk scheme.
But in the old scheme we did not notice the error whereas in the new scheme we
get an error in PostTyper that the type Iterable does not conform to its bound
`[X0] -> Iterable[X0]`.

With this commit, we change the scheme, so that eta-expansion depends on the
type parameters of a type itself, instead of the expected type.

We should investigate whether we can do a similar change for Scala2 classloading.