Make TypeAccumulators follow LazyRefs
TypeMaps do the same, so it is logical, and helps
prevent subtle errors as when we mispredicted whether
a RecType contains references that point to it.

Also, add normalizeHkApply to homogenize

Fixes some discrepancies in Tasty typing.

Also, homogenize skolem types

Skolem types are eliminated by pickling, so they should not appear
in the "before-pickling" output.