Revert "Generalize compareHkApply to types other than PolyParams"
[This was an unncessessary and wrong fix. The real problem was in findMember]

This seems to be required to make Typer compile, and it is logical
that we should do this. We still need to study how it fits in the
general context.

Also, it required neg test to be fixed. With the change, the error is
now reported earlier. (reverted from commit a6caf1ccf9a995e0ae3e1e78f40839afe3183f36)