Fix condition for lambda abstracting in Namer
The previous condition could make a (derived) type a * type
for a little while even though it had type parameters. This
loophole caused collection/generic/MapFactory.scala and with
it compile-stdlib to fail.
Refinement for knownHK for PolyParams

pos test t2082.scala shows that knownHK can be constructed before
the binder PolyType of a PolyParam is initialized.