Less eager removal of type parameters from constraint
Previously, a unified or otherwise instantiated type parameter
would be removed from the constraint, i.e. if it was the
last parameter of its polytype to be instantiated, the polytype
would be dropped. This is a potential problem since it means
that the alias `param = instance` is forgetten whereas we might
still need it in the same subtype test sequence.

The solution is to wait with cleaning up polytypes until
an enclosing subtype test has been completely unwound.

Also: Generalize constraint solving to GenericTypes (will be
needed later for handling existentials).