Implement phantom types.
In this commit phantom types are defined and implemented.
Only a minimal version of phantom term erasure is included.

Phantom types are in latices outside of the normal `Any`/`Nothing` type lattice.
A new phantom lattice can be defined by extending an `object` with `scala.Phantom`.
This trait defines synthetic members
* `protected final trait Any` where this `Any` does not extends `scala.Any`
* `protected final abstract class Nothing extends this.Any`
* `protected final def assume[P >: this.Nothing <: this.Any]: P`

A phantom lattice object can expose any of the phantom members (`Any`, `Nothing`, `assume`)
using an alias but it is not required.

Restriction on lattices:
* Types from different lattice can not be mixed using `&` or `|` types.
* Type parameters must be bounded by types in a single lattice.

Phantom type erasure (minimalistic):
* Type references `Phantom.Any` and `Phantom.Nothing` are erased to `dotty.runtime.ErasedPhantom`
* `Phantom.assume[P]` is erased to `null.asInstanceOf[P]`
* `Phantom` is erased to `dotty.runtime.ErasedPhantomLattice`, removing the members of `Phantom`