Quick-fix eclosingMethod attribute generation for t3048.scala
It's a funny interaction between:

    elim-by-name(and erasure specifically);
    lift-static;
    supercalls.

object E extends F2(new B {})

Here we have an anonymous class new B {} that looks like it is created
by erasure.

For some reason this class forgets the link to original anonymous class:
SymDenot(E$annon1).initial.phase == erasure. I guess this is a bug.

Additionally, the owner of E$annon1 is an anonymous method inside E,
that is inSuperCall and thus we have an anonymous nested class that
has enclosingClass be package. This class looks like a top-level
anonymous class breaking a lot of assumptions in shared backend
and taking multiple branches in unexpected ways.

I'm not sure that this is a proper fix. I assume there's a bigger bug
around, but I don't quite understand it right now and I need to work on
other stuff.

Making a quick fix to unblock @odersky.