SelectStatic: do not create blocks that are qualifier of select\apply
Blocks are not denoting trees(why aren't they?)
For now, I'm fixing this using a quick fix.
For future, it may make sense to discuss this on dotty meeting and
make blocks be a Denoting tree and return denotation of expo.

Another option is to move regularisation logic into tree transformers.