Refinements to auto tupling
There's a nasty interaction with auto-tupling and trying to insert an implicit
on the qualifier of a call. If the original call fails, we need to "undo" any
auto-tupling decisions in calls where an implicit is inserted on the qualifier.

Also: Needed to fix canAutoTuple test so that Scala2 feature is checked instead of dotty's.

Also: Drop features in dotty.language that duplicate those in scala.language.