Retracting special case depending on Ycheck.
I believe it's better to just bite the bullet and insert the cast.
If it should become a problem, we can think of a fallback, e.g. marking the
expression with a special tag, so that it does not get typechecked. But
for the moment I am not sure it is an issue at all.