Fix #1543: Add synthetic FunctionN types.
With this change it is possible to have create types for
functions of any arity. This ensures that the compiler will
not crash when creating FunProto when typing a method
application method applications with more than MaxFunctionArity
parameters. Note that in the language the functions are still
bound to MaxFunctionArity, but this is a first small step towards
supporting any function arity.